Creating audio utils
